+++ src/blueprints/telegram_bot/webhook/commands/upload.py
@@ -4,7 +4,9 @@ from typing import Union
 from flask import g, current_app
 
 from src.api import telegram
-from src.blueprints.telegram_bot._common import telegram_interface
+from src.blueprints.telegram_bot._common.telegram_interface import (
+    Message as TelegramMessage
+)
 from src.blueprints.telegram_bot._common.yandex_disk import (
     upload_file_with_url,
     YandexAPIRequestError,
@@ -19,6 +21,7 @@ from ._common.decorators import (
 from ._common.responses import (
     abort_command,
     cancel_command,
+    send_yandex_disk_error,
     AbortReason
 )
 
@@ -33,8 +36,10 @@ class MessageHealth:
         abort_reason: Union[AbortReason, None] = None
     ) -> None:
         """
-        :param ok: Message is valid for subsequent handling.
-        :param abort_reason: Reason of abort. `None` if `ok = True`.
+        :param ok:
+        Message is valid for subsequent handling.
+        :param abort_reason:
+        Reason of abort. `None` if `ok = True`.
         """
         self.ok = ok
         self.abort_reason = None
@@ -47,22 +52,25 @@ class AttachmentHandler(metaclass=ABCMeta):
     - most of attachments will be treated as files.
     - some of the not abstract class functions are common
     for most attachments. If you need specific logic in some
-    function, then override this function.
+    function, then override it.
     """
     def __init__(self) -> None:
         # Sended message to Telegram user.
-        # This message will be updated, rather
-        # than sending new message every time
-        self.sended_message: Union[
-            telegram_interface.Message,
-            None
-        ] = None
+        # This message will be updated, instead
+        # than sending new message every time again
+        self.sended_message: Union[TelegramMessage, None] = None
 
     @staticmethod
     @abstractmethod
     def handle(*args, **kwargs) -> None:
         """
         Starts uploading process.
+
+        - `*args`, `**kwargs` - arguments from dispatcher.
+
+        :raises:
+        Raises an error if any! So, this function should
+        be handled by top-function.
         """
         pass
 
@@ -70,7 +78,8 @@ class AttachmentHandler(metaclass=ABCMeta):
     @abstractmethod
     def telegram_action(self) -> str:
         """
-        :returns: Action type from
+        :returns:
+        Action type from
         https://core.telegram.org/bots/api/#sendchataction
         """
         pass
@@ -79,8 +88,9 @@ class AttachmentHandler(metaclass=ABCMeta):
     @abstractmethod
     def raw_data_key(self) -> str:
         """
-        :returns: Key in message, under this key
-        stored needed raw data. Example: 'audio'.
+        :returns:
+        Key in message, under this key stored needed raw data.
+        Example: `audio`.
         See https://core.telegram.org/bots/api#message
         """
         pass
@@ -89,42 +99,60 @@ class AttachmentHandler(metaclass=ABCMeta):
     @abstractmethod
     def raw_data_type(self) -> type:
         """
-        :returns: Expected type of raw data.
-        Example: 'dict'. `None` never should be returned!
+        :returns:
+        Expected type of raw data.
+        Example: `dict`.
+        `None` never should be returned!
+        """
+        pass
+
+    @abstractmethod
+    def create_help_message(self) -> str:
+        """
+        - supports HTML markup.
+
+        :returns:
+        Help message that will be sended to user
+        in case of some triggers (for example, when
+        there are no suitable data for handling).
         """
         pass
 
     def get_attachment(
         self,
-        message: telegram_interface.Message
+        message: TelegramMessage
     ) -> Union[dict, str, None]:
         """
-        :param message: Incoming Telegram message.
+        :param message:
+        Incoming Telegram message.
 
-        :returns: Attachment of message (photo object,
-        file object, audio object, etc.). If `None`, then
-        uploading should be aborted. If `dict`, it will
-        have `file_id` and `file_unique_id` properties.
-        If `str`, it should be assumed as direct file URL.
+        :returns:
+        Attachment of message (photo object, file object,
+        audio object, etc.). If `None`, then uploading should
+        be aborted. If `dict`, it will have `file_id` and
+        `file_unique_id` properties. If `str`, it should be
+        assumed as direct file URL.
         See https://core.telegram.org/bots/api/#available-types
         """
         return message.raw_data.get(self.raw_data_key)
 
     def check_message_health(
         self,
-        message: telegram_interface.Message
+        message: TelegramMessage
     ) -> MessageHealth:
         """
-        :param message: Incoming Telegram message.
+        :param message:
+        Incoming Telegram message.
 
-        :returns: See `MessageHealth` documentation.
+        :returns:
+        See `MessageHealth` documentation.
         Message should be handled by next operators only
         in case of `ok = true`.
         """
         health = MessageHealth(True)
         value = self.get_attachment(message)
 
-        if not (isinstance(value, self.raw_data_type)):
+        if not isinstance(value, self.raw_data_type):
             health.ok = False
             health.abort_reason = AbortReason.NO_SUITABLE_DATA
         elif (
@@ -148,34 +176,43 @@ class AttachmentHandler(metaclass=ABCMeta):
         file: Union[dict, None]
     ) -> str:
         """
-        :param attachment: Not `None` value from `self.get_attachment()`.
-        :param file: Representation of this attachment as a file on
-        Telegram servers. If `attachment` is `str`, then this will
-        be equal `None`. See https://core.telegram.org/bots/api/#file
+        :param attachment:
+        Not `None` value from `self.get_attachment()`.
+        :param file:
+        Representation of this attachment as a file on
+        Telegram servers. If `attachment` is `str`, then
+        this will be equal `None`.
+        See https://core.telegram.org/bots/api/#file
 
-        :returns: Name of file which will be uploaded.
+        :returns:
+        Name of file which will be uploaded.
         """
-        if (isinstance(attachment, str)):
+        if isinstance(attachment, str):
             return attachment
 
-        name = attachment.get("file_name") or file["file_unique_id"]
+        name = (
+            attachment.get("file_name") or
+            file["file_unique_id"]
+        )
         extension = self.get_mime_type(attachment)
 
-        if (extension):
+        if extension:
             name = f"{name}.{extension}"
 
         return name
 
     def get_mime_type(self, attachment: dict) -> str:
         """
-        :param attachment: `dict` result from `self.get_attachment()`.
+        :param attachment:
+        `dict` result from `self.get_attachment()`.
 
-        :returns: Empty string in case if `attachment` doesn't
-        have required key. Otherwise mime type of this attachment.
+        :returns:
+        Empty string in case if `attachment` doesn't have
+        required key. Otherwise mime type of this attachment.
         """
         result = ""
 
-        if ("mime_type" in attachment):
+        if "mime_type" in attachment:
             types = attachment["mime_type"].split("/")
             result = types[1]
 
@@ -185,57 +222,75 @@ class AttachmentHandler(metaclass=ABCMeta):
         """
         Checks if size of file exceeds limit size of upload.
 
-        :param file: `dict` value from `self.get_attachment()`.
+        :param file:
+        `dict` value from `self.get_attachment()`.
 
-        :returns: File size exceeds upload limit size.
+        :returns:
+        File size exceeds upload limit size.
         Always `False` if file size is unknown.
         """
         limit = current_app.config["TELEGRAM_API_MAX_FILE_SIZE"]
         size = limit
 
-        if ("file_size" in file):
+        if "file_size" in file:
             size = file["file_size"]
 
         return (size > limit)
 
     @yd_access_token_required
     @get_db_data
-    def upload(self) -> None:
+    def upload(self, *args, **kwargs) -> None:
         """
         Uploads an attachment.
+
+        `*args`, `**kwargs` - arguments from dispatcher.
         """
-        message = g.telegram_message
-        user = g.db_user
-        chat = g.db_chat
+        chat_id = kwargs.get(
+            "chat_id",
+            g.db_chat.telegram_id
+        )
+        message = kwargs.get(
+            "message",
+            g.telegram_message
+        )
         message_health = self.check_message_health(message)
 
-        if not (message_health.ok):
-            return abort_command(
-                chat.telegram_id,
-                message_health.abort_reason or AbortReason.UNKNOWN
+        if not message_health.ok:
+            reason = (
+                message_health.abort_reason or
+                AbortReason.UNKNOWN
             )
 
-        attachment = self.get_attachment(message)
+            if (reason == AbortReason.NO_SUITABLE_DATA):
+                return self.send_html_message(
+                    chat_id,
+                    self.create_help_message()
+                )
+            else:
+                return abort_command(chat_id, reason)
 
-        if (attachment is None):
-            return abort_command(
-                chat.telegram_id,
-                AbortReason.NO_SUITABLE_DATA
+        attachment = self.get_attachment(message)
+        data_is_empty = (attachment is None)
+
+        if data_is_empty:
+            return self.send_html_message(
+                chat_id,
+                self.create_help_message()
             )
 
         try:
             telegram.send_chat_action(
-                chat_id=chat.telegram_id,
+                chat_id=chat_id,
                 action=self.telegram_action
             )
         except Exception as error:
-            print(error)
-            return cancel_command(chat.telegram_id)
+            cancel_command(chat_id)
+            raise error
 
         download_url = None
         file = None
 
-        if (isinstance(attachment, str)):
+        if isinstance(attachment, str):
             download_url = attachment
         else:
             result = None
@@ -245,14 +300,15 @@ class AttachmentHandler(metaclass=ABCMeta):
                     file_id=attachment["file_id"]
                 )
             except Exception as error:
-                print(error)
-                return cancel_command(chat.telegram_id)
+                cancel_command(chat_id)
+                raise error
 
             file = result["content"]
             download_url = telegram.create_file_download_url(
                 file["file_path"]
             )
 
+        user = g.db_user
         file_name = self.create_file_name(attachment, file)
         user_access_token = user.yandex_disk_token.get_access_token()
         folder_path = current_app.config[
@@ -267,67 +323,99 @@ class AttachmentHandler(metaclass=ABCMeta):
                     file_name=file_name,
                     download_url=download_url
                 ):
-                    self.send_message(f"Status: {status}")
+                    self.reply_to_message(
+                        message.message_id,
+                        chat_id,
+                        f"Status: {status}"
+                    )
             except YandexAPICreateFolderError as error:
                 error_text = str(error) or (
                     "I can't create default upload folder "
-                    "due to an unknown Yandex error."
+                    "due to an unknown Yandex.Disk error."
                 )
 
-                return self.send_message(error_text)
+                return send_yandex_disk_error(
+                    chat_id,
+                    error_text,
+                    message.message_id
+                )
             except YandexAPIUploadFileError as error:
                 error_text = str(error) or (
                     "I can't upload this due "
-                    "to an unknown Yandex error."
+                    "to an unknown Yandex.Disk error."
                 )
 
-                return self.send_message(error_text)
+                return send_yandex_disk_error(
+                    chat_id,
+                    error_text,
+                    message.message_id
+                )
             except YandexAPIExceededNumberOfStatusChecksError:
                 error_text = (
                     "I can't track operation status of "
                     "this anymore. Perform manual checking."
                 )
 
-                return self.send_message(error_text)
-            except (YandexAPIRequestError, Exception) as error:
-                print(error)
-
-                if (self.sended_message is None):
-                    return cancel_command(
-                        chat.telegram_id,
+                return self.reply_to_message(
+                    message.message_id,
+                    chat_id,
+                    error_text
+                )
+            except Exception as error:
+                if self.sended_message is None:
+                    cancel_command(
+                        chat_id,
                         reply_to_message=message.message_id
                     )
                 else:
-                    return cancel_command(
-                        chat.telegram_id,
+                    cancel_command(
+                        chat_id,
                         edit_message=self.sended_message.message_id
                     )
 
+                raise error
+
         long_task()
 
-    def send_message(self, text: str) -> None:
+    def send_html_message(
+        self,
+        chat_id: int,
+        html_text: str
+    ) -> None:
         """
-        Sends message to Telegram user.
-
-        - if message already was sent, then sent message
-        will be updated with new text.
+        Sends HTML message to Telegram user.
         """
-        incoming_message = g.telegram_message
-        chat = g.db_chat
+        telegram.send_message(
+            chat_id=chat_id,
+            text=html_text,
+            parse_mode="HTML"
+        )
 
-        if (self.sended_message is None):
+    def reply_to_message(
+        self,
+        incoming_message_id: int,
+        chat_id: int,
+        text: str
+    ) -> None:
+        """
+        Sends reply message to Telegram user.
+
+        - if message already was sent, then sent
+        message will be updated with new text.
+        """
+        if self.sended_message is None:
             result = telegram.send_message(
-                reply_to_message_id=incoming_message.message_id,
-                chat_id=chat.telegram_id,
+                reply_to_message_id=incoming_message_id,
+                chat_id=chat_id,
                 text=text
             )
-            self.sended_message = telegram_interface.Message(
+            self.sended_message = TelegramMessage(
                 result["content"]
             )
         elif (text != self.sended_message.get_text()):
             telegram.edit_message_text(
                 message_id=self.sended_message.message_id,
-                chat_id=chat.telegram_id,
+                chat_id=chat_id,
                 text=text
             )
 
@@ -339,7 +427,7 @@ class PhotoHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = PhotoHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -354,13 +442,24 @@ class PhotoHandler(AttachmentHandler):
         # dict, not list, because we will select biggest photo
         return dict
 
-    def get_attachment(self, message: telegram_interface.Message):
+    def create_help_message(self):
+        return (
+            "Send a photos that you want to upload."
+            "\n\n"
+            "Note:"
+            "\n"
+            "- original name will be not saved"
+            "\n"
+            "- original quality and size will be decreased"
+        )
+
+    def get_attachment(self, message: TelegramMessage):
         photos = message.raw_data.get(self.raw_data_key, [])
         biggest_photo = None
         biggest_pixels_count = -1
 
         for photo in photos:
-            if (self.is_too_big_file(photo)):
+            if self.is_too_big_file(photo):
                 continue
 
             current_pixels_count = photo["width"] * photo["height"]
@@ -379,7 +478,7 @@ class FileHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = FileHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -393,6 +492,17 @@ class FileHandler(AttachmentHandler):
     def raw_data_type(self):
         return dict
 
+    def create_help_message(self):
+        return (
+            "Send a files that you want to upload."
+            "\n\n"
+            "Note:"
+            "\n"
+            "- original name will be saved"
+            "\n"
+            "- original quality and size will be saved"
+        )
+
     def get_mime_type(self, attachment):
         # file name already contains type
         return ""
@@ -405,7 +515,7 @@ class AudioHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = AudioHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -419,18 +529,31 @@ class AudioHandler(AttachmentHandler):
     def raw_data_type(self):
         return dict
 
+    def create_help_message(self):
+        return (
+            "Send a music that you want to upload."
+            "\n\n"
+            "Note:"
+            "\n"
+            "- original name will be saved"
+            "\n"
+            "- original quality and size will be saved"
+            "\n"
+            "- original type may be changed"
+        )
+
     def create_file_name(self, attachment, file):
         name = file["file_unique_id"]
 
-        if ("title" in attachment):
+        if "title" in attachment:
             name = attachment["title"]
 
-        if ("performer" in attachment):
+        if "performer" in attachment:
             name = f"{attachment['performer']} - {name}"
 
         extension = self.get_mime_type(attachment)
 
-        if (extension):
+        if extension:
             name = f"{name}.{extension}"
 
         return name
@@ -443,7 +566,7 @@ class VideoHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = VideoHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -457,6 +580,19 @@ class VideoHandler(AttachmentHandler):
     def raw_data_type(self):
         return dict
 
+    def create_help_message(self):
+        return (
+            "Send a video that you want to upload."
+            "\n\n"
+            "Note:"
+            "\n"
+            "- original name will be saved"
+            "\n"
+            "- original quality and size will be saved"
+            "\n"
+            "- original type may be changed"
+        )
+
 
 class VoiceHandler(AttachmentHandler):
     """
@@ -465,7 +601,7 @@ class VoiceHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = VoiceHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -479,6 +615,11 @@ class VoiceHandler(AttachmentHandler):
     def raw_data_type(self):
         return dict
 
+    def create_help_message(self):
+        return (
+            "Send a voice message that you want to upload."
+        )
+
 
 class URLHandler(AttachmentHandler):
     """
@@ -487,7 +628,7 @@ class URLHandler(AttachmentHandler):
     @staticmethod
     def handle(*args, **kwargs):
         handler = URLHandler()
-        handler.upload()
+        handler.upload(*args, **kwargs)
 
     @property
     def telegram_action(self):
@@ -501,7 +642,18 @@ class URLHandler(AttachmentHandler):
     def raw_data_type(self):
         return str
 
-    def get_attachment(self, message: telegram_interface.Message):
+    def create_help_message(self):
+        return (
+            "Send a direct URL to file that you want to upload."
+            "\n\n"
+            "Note:"
+            "\n"
+            "- original name from URL will be saved"
+            "\n"
+            "- original quality and size will be saved"
+        )
+
+    def get_attachment(self, message: TelegramMessage):
         return message.get_entity_value(self.raw_data_key)
 
     def create_file_name(self, attachment, file):