+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `/unpublish`: unpublishing of files or folders.
 - `/space`: getting of information about remaining Yandex.Disk space.
 - Handling of files that exceed file size limit in 20 MB. At the moment the bot will asnwer with warning message, not trying to upload that file. [#3](https://github.com/Amaimersion/yandex-disk-telegram-bot/issues/3)
+- Help messages for each upload command will be sended when there are no suitable input data.
 
 ### Changed
 
@@ -21,6 +22,7 @@
 
 - `/create_folder`: fixed a bug when bot could remove `/create_folder` occurrences from folder name.
 - `/create_folder`: fixed a bug when bot don't send any responses on invalid folder name.
+- Wrong information in help message for `/upload_video`.
 
 ## Project